Refactor: 게임 세부 설정 component로 분리
